refactor: 복잡한 로직을 훅으로 분리한다.
